# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Valhalla Core Stock API
  description: |
    API de autenticação, usuários, perfis, permissões e estoque.

    Operações autenticadas que alteram estado exigem validação CSRF.
    Após login ou refresh, a API também emite o cookie XSRF-TOKEN. Chamadas de um
    frontend na origem configurada podem usá-lo automaticamente; integrações que
    enviam o header X-XSRF-TOKEN continuam suportadas. GET /auth/csrf permanece
    disponível para clientes que precisem obter o valor explicitamente.
  version: v1
servers:
- url: http://localhost
  description: Generated server url
...
paths:
...
  /auth/refresh:
    post:
      tags:
      - Autenticação
      summary: Renovar autenticação
      description: Utiliza o refresh token armazenado no cookie para gerar novos tokens.
      operationId: refresh
      parameters:
      - name: refreshToken
        in: cookie
        description: Cookie HttpOnly com o refresh token.
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Autenticação renovada com sucesso.
          headers:
            Set-Cookie:
              description: Cookies HttpOnly accessToken e refreshToken são substituídos.
              style: simple
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespostaAutenticacaoDto"
        "401":
          description: "Refresh token ausente, inválido ou expirado."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - csrfTokenHeader: []
  /auth/logout:
    post:
      tags:
      - Autenticação
      summary: Realizar logout
      description: Invalida os tokens atuais e remove os cookies de autenticação.
      operationId: logout
      parameters:
      - name: refreshToken
        in: cookie
        description: Cookie HttpOnly com o refresh token.
        required: false
        schema:
          type: string
      - name: accessToken
        in: cookie
        description: Cookie HttpOnly com o access token.
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Logout realizado e cookies removidos.
          headers:
            Set-Cookie:
              description: Cookies accessToken e refreshToken são expirados.
              style: simple
      security:
      - csrfTokenHeader: []
  /auth/login:
    post:
      tags:
      - Autenticação
      summary: Realizar login
      description: Autentica o usuário e cria os cookies de acesso e renovação da
        sessão.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisicaoLoginDto"
        required: true
      responses:
        "200":
          description: Autenticação realizada com sucesso.
          headers:
            Set-Cookie:
              description: Cookies HttpOnly accessToken e refreshToken são criados.
              style: simple
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespostaAutenticacaoDto"
        "401":
          description: Usuário ou senha inválidos.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "400":
          description: Dados de entrada inválidos.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /auth/csrf:
    get:
      tags:
      - Autenticação
      summary: Obter token CSRF da sessão autenticada
      operationId: obterTokenCsrf
      responses:
        "200":
          description: Token CSRF retornado.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespostaCsrfTokenDto"
        "401":
          description: "Sessão ausente, inválida ou expirada."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
...
    RespostaAutenticacaoDto:
      type: object
      description: Confirmação de uma operação de autenticação
      properties:
        mensagem:
          type: string
          example: Autenticação realizada com sucesso
        usuario:
          type: string
          example: Administrador Inicial
    RequisicaoLoginDto:
      type: object
      properties:
        email:
          type: string
          format: email
          description: E-mail utilizado para realizar a autenticação
          example: usuario@exemplo.com
          minLength: 1
        senha:
          type: string
          description: Senha utilizada para realizar a autenticação
          example: Senha@123
          minLength: 1
      required:
      - email
      - senha
...
    RespostaCsrfTokenDto:
      type: object
      description: "Token CSRF de uma sessão autenticada, disponível também no cookie XSRF-TOKEN"
      properties:
        token:
          type: string
          example: YxBq8vlPoA1pP2dM0R2N9Q
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      description: Cookie HttpOnly criado por POST /auth/login.
      name: accessToken
      in: cookie
    csrfTokenHeader:
      type: apiKey
      description: "Obrigatório em POST, PUT, PATCH e DELETE autenticados."
      name: X-XSRF-TOKEN
      in: header